Refactor, update Amount attr from String to double
