+++ src/main/java/seedu/address/model/finance/logentry/DescriptionContainsKeywordsPredicate.java
@@ -1,0 +1,32 @@
+package seedu.address.model.finance.logentry;
+
+import java.util.List;
+import java.util.function.Predicate;
+
+import seedu.address.commons.util.StringUtil;
+
+
+/**
+ * Tests that a {@code LogEntry}'s {@code Description} matches any of the keywords given.
+ */
+public class DescriptionContainsKeywordsPredicate implements Predicate<LogEntry> {
+    private final List<String> keywords;
+
+    public DescriptionContainsKeywordsPredicate(List<String> keywords) {
+        this.keywords = keywords;
+    }
+
+    @Override
+    public boolean test(LogEntry logEntry) {
+        return keywords.stream()
+                .anyMatch(keyword -> StringUtil.containsWordIgnoreCase(logEntry.getDescription().value, keyword));
+    }
+
+    @Override
+    public boolean equals(Object other) {
+        return other == this // short circuit if same object
+                || (other instanceof DescriptionContainsKeywordsPredicate // instanceof handles nulls
+                && keywords.equals(((DescriptionContainsKeywordsPredicate) other).keywords)); // state check
+    }
+
+}
