+++ src/main/java/seedu/address/ui/finance/LogEntryCard.java
@@ -44,7 +44,7 @@ public class LogEntryCard extends UiPart<Region> {
         super(FXML);
         this.logEntry = logEntry;
         id.setText(displayedIndex + ". ");
-        amount.setText(logEntry.getAmount().amount);
+        amount.setText(logEntry.getAmount().toString());
         tDate.setText(logEntry.getTransactionDate().value);
         description.setText(logEntry.getDescription().value);
         tMethod.setText(logEntry.getTransactionMethod().value);
